Fix program version definition
The CFLAGS setting in the Makefile is ignore when compiling with
Buildroot since Buildroot sets the CFLAGS. This had the effect of
PROGRAM_VERSION not being defined and then anyone debugging erlinit
would see "unknown" as the version.

Since it is never good for the version to be unknown, make this a fatal
error when compiling.