Log to /dev/kmsg to avoid blocking
This fixes a situation where output was being sent to stderr and
blocking. This occurs in gadget mode where the output is going to the
virtual serial port, but the system on the other end doesn't have the
serial port open. The manifestation of this issue was that a reboot
would be delayed until the serial port was connected.

Log messages can be retreived from `dmesg` now.