Add support for Elixir's consolidated directory
This is needed to significantly speed up use of Elixir Protocols. Only
one directory is currently supported since that's also that should
exist. If more than one exists, we'll have to revisit since it's not
clear to me what order they should be added to Erlang's load path and
how we'd know.

Fixes #18.