Add /dev/rootdisk* helper symlinks
This is needed to sanely refer to the disk that provides the root
filesystem. Without it, hacks are needed or you need to start udevd to
create the /dev/disk/by-* directories and symlinks. None of these are
particularly convenient since the files aren't available as early on in
bootup as what erlinit can do.