Use custom test fixture for regression tests
This is a rewrite of the erlinit regression test framework to use a
shared library that intercepts and simulates C library calls so that the
erlinit can be exercised. This is similar to what fakechroot did, but
makes it possible to verify calls to mount, reboot, etc. that couldn't
be tested before. It also makes it possible to run the regression test
suite on OSX.