Restore sigmask for child proceses
To prevent a race condition, erlinit masks a set of signals until it's
ready to handle them. The signal mask gets inherited to child processes
and this may not be what they want or expect. This change restores the
signal mask to its original value.

This change fixes an issue where the application being launched by
erlinit was waiting for a SIGCHLD, but never received it since it was
masked.