Use erts version as specified by start_erl.data
Rather than scanning for erts directories, try the one in start_erl.data
first. If it doesn't exist, emit a warning and scan for available erts
directories.

In production Nerves systems there really should only be one erts
directory. However, it is possible to have two in development and this
ensures that you get the right one. Getting a warning on the specified
one doesn't exist seems useful. One could argue that this is
a fatal error. Since it hasn't been a fatal error in the past and making
it a fatal error could brick some Nerves deployments that can't recover
from bad builds if Erlang code doesn't run, I'm leaving it as a warning
for now.