Merge arguments to alternate-exec for sh -c
If the alterate-exec program invokes Erlang using "sh -c", then the
arguments need to be merged. Try to detect this by scanning for a call
to "exec" which one would tend to call for "sh -c". This makes erlinit
work with 'run_erl' which is nice since 'run_erl' is a standard
Erlang/OTP program with good docs.