Add --tty-options parameter for initializing tty
This makes it possible to initialize a tty via erlinit. Previously
erlinit relied on the Linux kernel to initialize the tty, but it would
only do this if specified via commandline paramters (e.g.,
console=ttyS0,115200n8). If the kernel didn't have this option set, it
can be inconvenient to enable and possibly confusing when no prints
come out.