Change rootdisk determination to use /sys/block
Erlinit creates `rootdisk` symlinks in /dev so that userland programs
can know which device is supplying the root filesystem. This is useful
since figuring this out other places can be inconvenient at best (i.e.,
run rdev and parse output). Often paths are hardcoded (/dev/mmcblk0p2)
and this makes it difficult to change boot devices or deal with cases
where Linux enumerates devices nondetermistically.

This commit replaces the heuristic implementation for figuring out which
device provides the root filesystem to one that queries /sys/block
files. The heuristic implementation guessed that the root device could
be determined by masking the device number of the root filesystem's
device.  This worked most of the time, but when it failed, the results
were really confusing.