Remove cr characters from log messages
These were an artifact of when logging was done to stderr. Ever since
the switch to use the kernel log, the carriage return characters have
been unnecessary.

In the unlikely event that the kernel log is unavailable, the logger
will write to stderr and the lack of cr characters will result in poor
formatting, but I believe that the first course of action when this
happens is to fix the kernel log. Also, the message won't be pretty, but
it will still convey the issue.