Add RELEASE_* environment vars to support runtime.exs
This makes it possible to use a `runtime.exs` file for runtime
configuration with Elixir. The following variables are added:

* `RELEASE_SYS_CONFIG`
* `RELEASE_ROOT`
* `RELEASE_TMP`

Without it, the release init script will terminate when the
`RELEASE_SYS_CONFIG` or `RELEASE_ROOT` environment variable can not be
found. Note that no support is here for copying the `sys.config` to a
writable location so that it can be updated for configuration changes
that require a VM restart. That is not supported.

The default Elixir release scripts specify more environment variables
and it's not clear whether they're needed or even if they make sense for
Nerves devices. Using `runtime.exs` with Nerves should be considered
experimental for now.