Protect against EINTR when running a command
The call to waitpid was missing the standard check for getting
interrupted by EINTR. This adds it. Amazingly, this doesn't seem to hit
on Linux, but is easy to reproduce on MacOS.

While making this change, I saw that the exit status was being passed
back raw rather than being extracted from the overall status. Due to how
bits are set in the overall status, this isn't an issue in practice yet,
but might be in the future.