Prefer ERTS found in release over system ERTS
This changes the default behavior of only using the system ERTS when
`--release-include-erts` isn't specified. The new behavior is to try the
release's ERTS first and then fallback to the system ERTS. The auto-ERTS
finding code isn't modified, but it also shouldn't be exercised much at
all any more.

The reasons for this change are:

1. Using the ERTS supplied in an OTP release is preferred everywhere I
   look these days. People will probably be surprised if it's not used.
2. Nerves currently uses ERTS from `/usr/lib/erlang`, but I'd like to
   change it to install ERTS with the release for consistency. Without
   this change, `--release-include-erts` would need to be added to all
   systems and it seems much easier to make erlinit "smarter".
3. OTP 25 uses `$ROOTDIR` more places. Running ERTS from the release
   gets rid of some confusion on where it needs to point. The confusion
   could be mine, but keeping ERTS with the release removes the issue.