Support relative paths to ERTS-supplied binaries
When the ERTS directory started being included in the OTP release (a
good thing), running `run_erl` as an alternate-exec program became
brittle. The reason is that the ERTS directory includes a version number
so you have to update the path to `run_erl` every Erlang/OTP update.

This makes it possible to pass a relative path to the alternate-exec
option and let `erlinit` make it point to the ERTS supplied bin
directory.