Switch from vt100 to xterm to avoid output delays
The vt100 setting turned out to not be harmless since it would
frequently cause `$<50>` sequences to be output. These tell the terminal
to delay 50 ms. Not all terminals support the output delay sequence so
these sequences would be user-visible.

OTP 26's new tty implementation properly follows terminfo so these
sequences were getting sent. This PR switches the `TERM` environment
variable to `xterm-256color`. The `xterm` terminals are pretty generic
and almost anything supports 256 colors these days. This also was the
default for the Linux and Macs I was using.