Reseed the rng with seedrng
This pulls in seedrng and invokes it on boot and shutdown to save and
restore random number seeds.

Seedrng can be found at https://git.zx2c4.com/seedrng/. Minimal changes
have been made to it to output messages consistent with erlinit. Errors
are ignored, so systems with read-only filesystems or filesystem errors
won't fail.