Extend the printf pattern for hostnames for left trimming
This provides more flexibility in specifying hostname patterns that
include all or part of the board's serial number. Before this change,
"%s" and "%.4s" (4 is the number of characters to use of the serial
number) were supported. After this change, the printf pattern also
supports "%-.4s" to use the rightmost characters in the board's serial
number. This is an extension to printf. I found one reference to someone
else doing this based on web searches and it was the only hit, so it
doesn't look like there's any precedence to follow.

While making this change, the custom sprintf was dumbed down to only
support these escapes and nothing else. This might prevent a crash.