Move a strtok call to avoid unnecessary invocation
This fixes what an unnecessary call to strtok when there's no alternate
exec being passed. It turns out that it's harmless since strtok handles
the NULL being passed in and the result is unused. However, it was
sloppy and identified by clang's sanitize feature. This should prevent
it from being flagged in the future.

Thanks to @jbnoblot for reporting this.