Add --core-pattern and document core dump config
If Erlang crashes right away, it's critical to get a core dump to see
what happened. The default core location wasn't in a writable location
so core dumps were lost. This adds a core pattern option to set where
the dumps should be written and documents the recipe for enabling them.