Fix umount parse issue with long options
Previously mounts with >127 characters of options would fail to parse
the options field. Due to the way the code was implemented, this would
still unmount everything, but sometimes it would try to unmount the
options which would obviously fail. If you were watching logs, it would
look really weird.

Side benefit is that this change reduces the stack usage quite a bit.