Support arbitrary baud rates
This has a couple uses cases. Some devices, notably Rockchip ones,
typically run their console UART at a high speed. While this is
optional, it's nice to match the default so that bootloader messages are
readable without temporarily switching rates. The second use case is to
bump the baud rate as high as possible so that using Erlang's peer via
the UART performs well.