Refactor all logging to fine tune priorities
This should be an invisible change to users who do not turn on verbose
logging. For those who do, the log messages are labelled better for
which ones are informational, warnings and errors.

This also adds compiler checking on printf format strings which ended up
catching two format string bugs.