Log to /dev/pmsg if available
This makes it possible to see what's happening before the Erlang VM
starts and after Erlang exits.

It only works if the Linux pstore driver is available.

With debug UART access, it was possible to get this information. If you
didn't have UART access, it was near impossible since these areas aren't
logged. Logging to pmsg at least at the level that erlinit does is
cheap, so it's enabled all the time since you usually don't know ahead
of time when there's going to be an unexpected reboot.